requirements.txt
requests>=2.8.1
urllib3>=2.2.2
pydantic[email]>=2.8.2
fastmcp>=2.13.0.2
msgraph-sdk>=1.0.0
azure-identity>=1.15.0
